cmd/genie: store and restore hook asm
This allows functions of the original binary to be hooked and invoked
more than once. Prior to this commit, a function hooked by genie would
have its original assembly restored, and as such the hook would no
longer be active. Now the hook (jump to genie generated function) is
stored before restoring the original asm of the hooked function, and
then once the original function has finished execution and returns
control to the genie generated function, the hook is restored.